Fixed showing message in push notification
